requirements.txt
beautifulsoup4==4.12.3
dnspython==2.6.1
dnsrecon==0.10.1
google==3.0.0
lxml==5.2.2
markdown-it-py==3.0.0
mdurl==0.1.2
netaddr==1.3.0
oxdork==3.2.0
Pygments==2.18.0
PyYAML==6.0.1
rich==13.7.1
soupsieve==2.5
tqdm==4.66.4
